Upgrade E2E tests to Core Tools v4.0.7317 (#88)
While trying to upgrade to PowerShell 7.4 and .NET 8, I discovered that
Core Tools builds are not published to the same internal location
anymore. As a result, the E2E tests were using a very old version of
Core Tools, which led to test failures.

This PR makes the E2E tests use Core Tools v4.0.7317 (the latest
officially published version), and it also removes the
`UseCoreToolsBuildFromIntegrationTests` support.

In the future, we may want to find another way of retrieving the latest
version, especially for the purposes of internal integration tests.
However, I believe public tests on GitHub should rely on officially
published Core Tools versions, and we should upgrade this dependency the
same way as we upgrade other dependencies: with an explicit version
bump.